refactor(daemon): move wardnetd-api dns_log_ws inline tests to api/tests/
Move the inline mod unit_tests out of api/dns_log_ws.rs into
src/api/tests/dns_log_ws.rs, wired via mod dns_log_ws; in
src/api/tests/mod.rs — matching the sibling logs_ws handler.
ClientFilter and its methods were already pub(crate), so no visibility
changes were needed. Pure move.

Part of #847.
